+++ user-ui/src/components/Ipregister/Ipregister.jsx
@@ -54,11 +54,11 @@ const Ipregister = () => {
     </div>
 
     <div>
-        <a class="arrow mb-4 bg-gradient-to-r from-black via-gray-300" onClick={routeChange}>Back</a>
+        <a class="arrow mb-6 bg-gradient-to-r from-black via-gray-300" onClick={routeChange}>Back</a>
     </div>
 
   <div data-testid="show"
-    className='container contain-1 text-gray-600 font-serif flex justify-between'>
+    className='regcont contain-1 text-gray-600 font-serif flex justify-between'>
       <form className='ipform px-5 mx-20 my-16 bg-white flex'>
        <div className='fields'>
         <div className='mb-6 py-3 text-black'><h1>Register IP</h1></div>
