+++ user-ui/src/css/Style.css
@@ -1,6 +1,7 @@
-.container {
+.regcont{
     background-color: rgba(231, 226, 226, 0.658);
     margin-bottom: 10rem;
+    /* margin-top: -1rem; */
 }
 .ipform,
 .bidform {
